code optimization, ENS changes, connection handling
