+++ src/components/Referrals/InviteCreators.tsx
@@ -12,12 +12,13 @@ import { fetchInviteeDetails } from "./InviteUtils";
 import RewardButton from "../ClaimReward/RewardButton";
 import ConnectWalletWithENS from "../ConnectWallet/ConnectWalletWithENS";
 import Link from "next/link";
+import { useConnection } from "@/app/hooks/useConnection";
 
 function InviteCreators({ userAddress }: { userAddress: any }) {
   const [copied, setCopied] = useState(false);
   const [inviteeDetails, setInviteeDetails] = useState<any>();
-  const [isPageLoading, setIsPageLoading] = useState(true);
-  // const [formattedAddr, setFormattedAddr] = useState("");
+  const { isConnected, isLoading, isSessionLoading, isPageLoading, isReady } =
+    useConnection();
 
   const handleCopy = () => {
     navigator.clipboard.writeText(`${BASE_URL}invite/${userAddress}`);
@@ -64,135 +65,118 @@ ${decodeURIComponent(urlEncoded)}
     const fetchData = async () => {
       const details = await fetchInviteeDetails(userAddress);
       setInviteeDetails(details);
-      setIsPageLoading(false);
+      // setIsPageLoading(false);
     };
     fetchData();
   }, []);
 
   return (
     <>
-      {isPageLoading ? (
-        <div className="flex items-center justify-center top-10 h-screen">
-          <Oval
-            visible={true}
-            height="40"
-            width="40"
-            color="#0500FF"
-            secondaryColor="#cdccff"
-            ariaLabel="oval-loading"
-          />
+      <div className="w-full flex justify-end pt-2 xs:pt-4 sm:pt-6 px-4 md:px-6 lg:px-14">
+        <div className="flex gap-1 xs:gap-2 items-center">
+          <RewardButton />
+          <ConnectWalletWithENS />
         </div>
-      ) : (
-        <>
-          <div className="w-full flex justify-end pt-2 xs:pt-4 sm:pt-6 px-4 md:px-6 lg:px-14">
-            <div className="flex gap-1 xs:gap-2 items-center">
-              <RewardButton />
-              <ConnectWalletWithENS />
+      </div>
+      <div className="p-8 bg-white">
+        <div className="text-2xl md:text-3xl font-semibold text-gray-900 mb-4">
+          Hi{" "}
+          {inviteeDetails?.ensName ||
+            inviteeDetails?.displayName ||
+            inviteeDetails?.formattedAddr}
+          , ready to earn ETH by inviting creators?
+        </div>
+        <div className="md:text-lg text-gray-700 mb-6">
+          Invite creators to Chora Club and earn ETH every time they mint! The
+          more you bring, the more you earn!
+        </div>
+        <div className="mb-8">
+          <Link
+            href={"https://docs.chora.club/earn-rewards/create-referral-reward"}
+            target="_blank"
+            className="text-blue-600 hover:text-blue-800 hover:underline transition-colors duration-300"
+          >
+            Learn more about referrals
+          </Link>
+        </div>
+
+        <div className="md:w-[70%] mx-auto">
+          <div className="p-4 sm:p-6 bg-gradient-to-r from-blue-50 via-blue-100 to-blue-50 text-blue-900 rounded-md flex flex-col items-center mb-8 w-full mx-auto">
+            <div className="flex rounded-full bg-white mt-2 sm:mt-3 mb-4 sm:mb-6">
+              <Image
+                src={
+                  inviteeDetails?.ensAvatar ||
+                  (inviteeDetails?.displayImage &&
+                    `https://gateway.lighthouse.storage/ipfs/${inviteeDetails.displayImage}`) ||
+                  user
+                }
+                width={100}
+                height={100}
+                alt="Invitee avatar"
+                className="h-20 w-20 sm:h-28 sm:w-28 object-cover rounded-full"
+              />
             </div>
-          </div>
-          <div className="p-8 bg-white">
-            <div className="text-2xl md:text-3xl font-semibold text-gray-900 mb-4">
-              Hi{" "}
+            <div className="font-semibold text-center mb-1 text-sm sm:text-base">
+              You&apos;ve been invited to create on Chora Club by
+            </div>
+            <div className="font-bold text-base sm:text-lg mb-2 text-center break-words max-w-full px-2">
               {inviteeDetails?.ensName ||
                 inviteeDetails?.displayName ||
                 inviteeDetails?.formattedAddr}
-              , ready to earn ETH by inviting creators?
             </div>
-            <div className="md:text-lg text-gray-700 mb-6">
-              Invite creators to Chora Club and earn ETH every time they mint!
-              The more you bring, the more you earn!
-            </div>
-            <div className="mb-8">
-              <Link
-                href={
-                  "https://docs.chora.club/earn-rewards/create-referral-reward"
-                }
-                target="_blank"
-                className="transition-colors"
-              >
-                Learn more about referrals
-              </Link>
-            </div>
-
-            <div className="md:w-[70%] mx-auto">
-              <div className="p-4 sm:p-6 bg-gradient-to-r from-blue-50 via-blue-100 to-blue-50 text-blue-900 rounded-md flex flex-col items-center mb-8 w-full mx-auto">
-                <div className="flex rounded-full bg-white mt-2 sm:mt-3 mb-4 sm:mb-6">
-                  <Image
-                    src={
-                      inviteeDetails?.ensAvatar ||
-                      (inviteeDetails?.displayImage &&
-                        `https://gateway.lighthouse.storage/ipfs/${inviteeDetails.displayImage}`) ||
-                      user
-                    }
-                    width={100}
-                    height={100}
-                    alt="Invitee avatar"
-                    className="h-20 w-20 sm:h-28 sm:w-28 object-cover rounded-full"
-                  />
-                </div>
-                <div className="font-semibold text-center mb-1 text-sm sm:text-base">
-                  You&apos;ve been invited to create on Chora Club by
-                </div>
-                <div className="font-bold text-base sm:text-lg mb-2 text-center break-words max-w-full px-2">
-                  {inviteeDetails?.ensName ||
-                    inviteeDetails?.displayName ||
-                    inviteeDetails?.formattedAddr}
-                </div>
-                <div className="mt-2 sm:mt-4">
-                  <Image
-                    src={logo}
-                    alt="Chora Club logo"
-                    width={40}
-                    height={40}
-                    className="w-10 h-10 sm:w-12 sm:h-12"
-                  />
-                </div>
-              </div>
-
-              <div className="flex flex-col w-full sm:flex-row items-center mb-8 gap-3">
-                <div className="w-full sm:flex-grow p-2 border border-gray-300 rounded-md overflow-hidden bg-gray-50 text-gray-700 cursor-pointer focus:outline-none truncate">
-                  {BASE_URL}invite/{userAddress}
-                </div>
-                <button
-                  className="w-full sm:w-auto bg-black text-white px-5 py-2 rounded-md hover:bg-gray-800 transition-colors duration-200"
-                  onClick={handleCopy}
-                >
-                  {copied ? "Copied!" : "Copy"}
-                </button>
-              </div>
-
-              <div className="flex gap-4 justify-center items-center mb-20">
-                <div
-                  className="bg-black rounded-full size-[40px] flex justify-center items-center cursor-pointer"
-                  onClick={shareOnTwitter}
-                >
-                  <RiTwitterXLine className="text-white bg-black size-6" />
-                </div>
-                <div onClick={shareOnFarcaster}>
-                  <SiFarcaster className="bg-white text-[#8a63d2] rounded-full size-[40px] cursor-pointer" />
-                </div>
-                <div onClick={shareOnTelegram}>
-                  <FaTelegram className="text-[#1d98dc] bg-white size-[40px] cursor-pointer" />
-                </div>
-              </div>
-
-              <div className="bg-gray-50 p-3 sm:p-4 rounded-md flex justify-between items-center flex-wrap gap-2">
-                <div className="flex items-center flex-wrap gap-2">
-                  <div className="text-xs sm:text-sm md:text-base text-gray-800">
-                    Total earnings from referrals
-                  </div>
-                  <button className="bg-black text-white text-xs sm:text-sm font-semibold px-2 sm:px-3 py-1 rounded-full hover:bg-gray-800 transition-colors">
-                    Claim
-                  </button>
-                </div>
-                <div className="text-blue-600 font-bold text-sm sm:text-base md:text-lg">
-                  {0} ETH
-                </div>
-              </div>
+            <div className="mt-2 sm:mt-4">
+              <Image
+                src={logo}
+                alt="Chora Club logo"
+                width={40}
+                height={40}
+                className="w-10 h-10 sm:w-12 sm:h-12"
+              />
             </div>
           </div>
-        </>
-      )}
+
+          <div className="flex flex-col w-full sm:flex-row items-center mb-8 gap-3">
+            <div className="w-full sm:flex-grow p-2 border border-gray-300 rounded-md overflow-hidden bg-gray-50 text-gray-700 cursor-pointer focus:outline-none truncate">
+              {BASE_URL}invite/{userAddress}
+            </div>
+            <button
+              className="w-full sm:w-auto bg-black text-white px-5 py-2 rounded-md hover:bg-gray-800 transition-colors duration-200"
+              onClick={handleCopy}
+            >
+              {copied ? "Copied!" : "Copy"}
+            </button>
+          </div>
+
+          <div className="flex gap-4 justify-center items-center mb-20">
+            <div
+              className="bg-black rounded-full size-[40px] flex justify-center items-center cursor-pointer"
+              onClick={shareOnTwitter}
+            >
+              <RiTwitterXLine className="text-white bg-black size-6" />
+            </div>
+            <div onClick={shareOnFarcaster}>
+              <SiFarcaster className="bg-white text-[#8a63d2] rounded-full size-[40px] cursor-pointer" />
+            </div>
+            <div onClick={shareOnTelegram}>
+              <FaTelegram className="text-[#1d98dc] bg-white size-[40px] cursor-pointer" />
+            </div>
+          </div>
+
+          <div className="bg-gray-50 p-3 sm:p-4 rounded-md flex justify-between items-center flex-wrap gap-2">
+            <div className="flex items-center flex-wrap gap-2">
+              <div className="text-xs sm:text-sm md:text-base text-gray-800">
+                Total earnings from referrals
+              </div>
+              <button className="bg-black text-white text-xs sm:text-sm font-semibold px-2 sm:px-3 py-1 rounded-full hover:bg-gray-800 transition-colors">
+                Claim
+              </button>
+            </div>
+            <div className="text-blue-600 font-bold text-sm sm:text-base md:text-lg">
+              {0} ETH
+            </div>
+          </div>
+        </div>
+      </div>
     </>
   );
 }
