+++ src/utils/ENSUtils.ts
@@ -3,7 +3,7 @@ import { cache } from "react";
 import { truncateAddress } from "./text";
 import { ethers } from "ethers";
 import { getEnsAvatar } from "@wagmi/core";
-import { getEnsName as ens } from "@wagmi/core";
+import { getEnsName } from "@wagmi/core";
 import { config } from "./config";
 import { normalize } from "viem/ens";
 import { mainnet } from "@wagmi/core/chains";
@@ -54,7 +54,7 @@ export async function getMetaAddressOrEnsName(
   //   `https://api.karmahq.xyz/api/dao/find-delegate?dao=${daoName}&user=${address}`
   // );
   // const details = await res.json();
-  const fetchedEnsName = await getEnsName(address);
+  const fetchedEnsName = await fetchEnsName(address);
   // const karmaEnsName = details.data.delegate?.ensName;
   const ensName = fetchedEnsName?.ensName;
   return ensName ? ensName : truncateAddress(address);
@@ -64,24 +64,30 @@ export async function getMetaProfileImage() {}
 
 export async function fetchEnsAvatar(address: any) {
   try {
-    const ensName = await ens(config, {
+    const ensName = await getEnsName(config, {
       address,
       chainId: mainnet.id,
     });
     const avatar = await getEnsAvatar(config, {
+      assetGatewayUrls: {
+        ipfs: "https://cloudflare-ipfs.com",
+      },
+      gatewayUrls: ["https://cloudflare-ipfs.com"],
       name: normalize(ensName?.toString() || ""),
       chainId: mainnet.id,
+      // universalResolverAddress: "0x74E20Bd2A1fE0cdbe45b9A1d89cb7e0a45b36376",
     });
     return { avatar, ensName };
   } catch (error) {
+    const truncatedAddress = truncateAddress(address);
     console.error(`Error fetching ENS details for address ${address}:`, error);
-    return null;
+    return { avatar: null, ensName: truncatedAddress };
   }
 }
 
-export async function getEnsName(address: any) {
+export async function fetchEnsName(address: any) {
   try {
-    const ensName = await ens(config, {
+    const ensName = await getEnsName(config, {
       address,
       chainId: mainnet.id,
     });
@@ -93,6 +99,8 @@ export async function getEnsName(address: any) {
 
     return { ensNameOrAddress, ensName };
   } catch (e) {
-    console.log("Error: ", e);
+    console.log("Error in fetchEnsName ", e);
+    const truncatedAddress = truncateAddress(address);
+    return { ensNameOrAddress: truncatedAddress, ensName: truncatedAddress };
   }
 }
